# WIP: internal/traceprof: make endpoint call counting efficient