@darccio 2025-03-10
@nsrip-dd As part of our upcoming release, we are reviewing all in-flight PRs. Please consider closing it or rebasing to `v2-dev` if you want to merge it for the next release after `v2.0.0`.