@nsrip-dd 2025-04-15
I'm not going to finish this. @felixge please feel free to take this over, though I would be hesitant to sign off on using this `xsync` library without further review. Is the library's implementation correct? Well tested? My instincts are telling me that the concurrency stuff in that library is quite subtle and I don't want to risk introducing painful-to-diagnose concurrency bugs just to squeeze out a bit of performance.